backend/requirements.txt
aniso8601==8.0.0
astroid==2.4.2
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
PyJWT==1.4.2
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
